Refactor MarginRequirementEntry to it's own file
